feat: refactor ProfileTileHeader import paths to use common directory
