+++ src/components/ai-matches/MatchPersonalInformation.tsx
@@ -1,6 +1,6 @@
 import { YStack, ViewProps } from 'tamagui';
 import { ProfileItem } from '../profile/ProfileItem';
-import { ProfileTileHeader } from '../profile/ProfileTileHeader';
+import { ProfileTileHeader } from '../common/ProfileTileHeader';
 import { genders } from '@/src/resources/gender';
 import { cities, states } from '@/src/resources/city-state';
 import { formatDateString } from '@/src/utils/dateFormatter';
