+++ src/components/profile/PersonalInformation.tsx
@@ -1,7 +1,7 @@
 import { User } from '@/src/models/User';
 import { YStack, ViewProps } from 'tamagui';
 import { ProfileItem } from './ProfileItem';
-import { ProfileTileHeader } from './ProfileTileHeader';
+import { ProfileTileHeader } from '../common/ProfileTileHeader';
 import { genders } from '@/src/resources/gender';
 import { cities, states } from '@/src/resources/city-state';
